// src/interfaces/customs/fics/GroundsBlockingInvestmentPlan.ts
export interface IGroundsBlockingInvestmentPlanTable {
  id: number
  code: unknown
  description: unknown
  status_id: unknown
  created_at: unknown
  updated_at: unknown
}

export interface IGroundsBlockingInvestmentPlanItemList {
  id?: number | null
  code?: number | null
  description?: string
  status_id?: string
  created_at?: string
  updated_at?: string | null
}